remmina: fix not relaunching
remmina would not relaunch because the application
would never close, but rather would continue to
run in the background.

The --no-tray-icon flag was added to the
connection command, which prevents the application
from continuing to run in the background, which
would cause the launch-work-rdp service to no
longer respond to new app.rdp files until the main
process was terminated.