rclone: persist rclone.conf items
Previously, management of `rclone.conf` by
home-manger was all-or-nothing. Now, with the
`rclone.remote.<remote>.persist` option, a list of
configuration items can be specified that should
be persisted across rclone runs.

This is most useful for managing access tokens
out-of-band from home-manager, while still
allowing home-manager to manage other items in the
`rclone.conf` file.